requirements.txt
altgraph==0.17
pefile==2019.4.18
numpy>=1.23.5
opencv-python==4.11.0.86
PyInstaller==5.13.1
pyserial==3.4
PySimpleGUI==5.0.8.2
python-chess==0.30.1
gTTS==2.3.1
#pygame==2.1.3
